Added enum for test sequence
