+++ src/main/java/com/mongodb/mongopush/constants/MongoPushConstants.java
@@ -16,8 +16,8 @@ public class MongoPushConstants {
 	public static final String DOT_LOG = ".log";
 	public static final String POCDRIVER_BINARY_DEFAULT_PATH = "pocdriver/binary/POCDriver.jar";
 	
-	public static final String TEST_SEQUENCE = "testSequence";
-	public static final String INCLUDE_OPTION = "includeOption";
+	public static final String TEST_SEQUENCE_NAME = "testSequenceName";
+	public static final String INCLUDE_OPTIONS = "includeOptions";
 	public static final String POC_DRIVER_ARGUMENTS = "pocdriveArguments";
 	
 	public static final String PUSH = "push";
