+++ src/test/resources/mongopush/pushDataTests
@@ -1,5 +1,5 @@
-ExecutePOCDriver->InitialDataInserted->ShutDownPOCDriver->ExecuteMongoPushModeData->InitialSyncCompleted->ShutDownMongoPush
-ExecutePOCDriver->InitialDataInserted->ShutDownPOCDriver->ExecuteMongoPushModeDataOnly->InitialSyncCompleted->ShutDownMongoPush
-ExecutePOCDriver->InitialDataInserted->ShutDownPOCDriver->ExecuteMongoPushModeDataOnly->InitialSyncCompleted->ShutDownMongoPush->ExecuteMongoPushModeVerify->VerificationTaskCompleted->ShutDownMongoPush
-ExecutePOCDriver->InitialDataInserted->ExecuteMongoPushModeData->InitialSyncCompleted->ShutDownPOCDriver->OplogStreamingCompleted->ShutDownMongoPush
-ExecutePOCDriver->InitialDataInserted->ExecuteMongoPushModeDataOnly->InitialSyncCompleted->ShutDownMongoPush->ShutDownPOCDriver->ExecuteMongoPushModeVerify->VerificationTaskFailed->ShutDownMongoPush
+POC_DRIVER_MONGO_PUSH_DATA_SYNC
+POC_DRIVER_MONGO_PUSH_DATA_ONLY_SYNC
+POC_DRIVER_MONGO_PUSH_DATA_ONLY_SYNC_VERIFY
+POC_DRIVER_MONGO_PUSH_DATA_SYNC_OPLOG
+POC_DRIVER_MONGO_PUSH_DATA_ONLY_SYNC_VERIFY_FAILED
