+++ src/test/resources/mongopush/tests/dataAndFilterTest01
@@ -1,5 +1,5 @@
 {
-	"testSequence": "ExecutePOCDriver->InitialDataInserted->ShutDownPOCDriver->ExecuteMongoPushModeData->InitialSyncCompleted->ShutDownMongoPush",
-	"includeOption": [{"namespace":"POCDB.POCCOLL","filter":{"fld0": {"$gt":500000}}}],
+	"testSequenceName": "POC_DRIVER_MONGO_PUSH_DATA_SYNC",
+	"includeOptions": [{"namespace":"POCDB.POCCOLL","filter":{"fld0": {"$gt":500000}}}],
 	"pocdriveArguments" : "-k 2 -i 1 -u 1 -b 2"
 }
