+++ src/test/resources/mongopush/tests/dataAndFilterTest02
@@ -1,4 +1,4 @@
 {
-	"testSequence": "ExecutePOCDriver->InitialDataInserted->ShutDownPOCDriver->ExecuteMongoPushModeDataOnly->InitialSyncCompleted->ShutDownMongoPush",
-	"includeOption": [{"namespace":"POCDB.POCCOLL","filter":{"fld0": {"$gt":500000}}}]
+	"testSequenceName": "POC_DRIVER_MONGO_PUSH_DATA_ONLY_SYNC",
+	"includeOptions": [{"namespace":"POCDB.POCCOLL","filter":{"fld0": {"$gt":500000}}}]
 }
